requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
pymysql
python-dotenv
